+++ frontend/src/components/Footer.stories.tsx
@@ -1,7 +1,7 @@
 import { createStore } from '@reduxjs/toolkit';
 import { Meta, StoryFn } from '@storybook/react';
 import { Provider } from 'react-redux';
-import { MemoryRouter } from 'react-router-dom';
+import { MemoryRouter } from 'react-router';
 
 import FooterComponent from './Footer';