+++ frontend/src/components/Groups/GroupCharts/VersionCountTimeline.tsx
@@ -2,8 +2,8 @@ import { Theme } from '@mui/material';
 import Box from '@mui/material/Box';
 import Chip from '@mui/material/Chip';
 import Grid from '@mui/material/Grid';
-import Typography from '@mui/material/Typography';
 import { useTheme } from '@mui/material/styles';
+import Typography from '@mui/material/Typography';
 import React from 'react';
 import semver from 'semver';