+++ frontend/src/components/layouts/PageNotFoundLayout/PageNotFoundLayout.tsx
@@ -1,6 +1,6 @@
 import { Grid, Typography } from '@mui/material';
 import { Trans, useTranslation } from 'react-i18next';
-import { Link } from 'react-router-dom';
+import { Link } from 'react-router';
 
 import pageNotFoundGraphic from './404.svg';